+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * `2024-03-04 13:54:55` : 
     - Renamed `X_PARAM` and `X_ALIAS` to `VENDOR_PARAM` and `VENDOR_ALIAS`
     - Make sure BCLD always reads the [BCLD_VERBOSE](./script/startup.sh) parameter first
+    - Renamed methods in [BCLD Vendor script](./script/bcld_vendor.sh)
 * `2024-02-27 17:32:47` : Fixed NSSDB permissions in [BCLD Vendor script](./script/bcld_vendor.sh)
 * `2024-02-26 14:49:36` : [OAC-836] BCLD now executes [HashGen](./tools/HashGen.sh) automatically as part of the new [pre-commit](.git/hooks/pre-commit) hook
 * `2024-02-26 12:46:52` : Converted ShellCheck report to Markdown format